Some changes in file formatting and flash-write malloc

+++ dev/flash.c
@@ -10,12 +10,12 @@
  *
  * version: 1.0
  *
- * description: 
+ * description:
  *
  * dependencies:
  *
  * references:
- * 
+ *
  *==============================================================================
  * GNU LESSER GENERAL PUBLIC LICENSE
  *==============================================================================
@@ -32,10 +32,10 @@
  * last changes:
  *    2013-09-25   Theodor Stana     t.stana@cern.ch     File created
  *==============================================================================
- * TODO: - 
+ * TODO: -
  *==============================================================================
  */
- 
+
 #include "flash.h"
 #include "types.h"
 #include "syscon.h"
@@ -51,30 +51,27 @@
 /*
  * Bit-bang SPI transfer function
  */
-static uint8_t bbspi_transfer(uint8_t cspin, uint8_t val)
+static uint8_t bbspi_transfer(int cspin, uint8_t val)
 {
-  uint8_t i, retval = 0;
-  gpio_out(GPIO_SPI_NCS, cspin);
-  for (i = 0; i < 8; i++)
-  {
-    gpio_out(GPIO_SPI_SCLK, 0);
-    if (val & 0x80)
-    {
-      gpio_out(GPIO_SPI_MOSI, 1);
-    }
-    else
-    {
-      gpio_out(GPIO_SPI_MOSI, 0);
-    }
-    gpio_out(GPIO_SPI_SCLK, 1);
-    retval <<= 1;
-    retval |= gpio_in(GPIO_SPI_MISO);
-    val <<= 1;
-  }
+	int i, retval = 0;
+	gpio_out(GPIO_SPI_NCS, cspin);
+	for (i = 0; i < 8; i++) {
+		gpio_out(GPIO_SPI_SCLK, 0);
+		if (val & 0x80) {
+			gpio_out(GPIO_SPI_MOSI, 1);
+		}
+		else {
+			gpio_out(GPIO_SPI_MOSI, 0);
+		}
+		gpio_out(GPIO_SPI_SCLK, 1);
+		retval <<= 1;
+		retval |= gpio_in(GPIO_SPI_MISO);
+		val <<= 1;
+	}
 
-  gpio_out(GPIO_SPI_SCLK, 0);
+	gpio_out(GPIO_SPI_SCLK, 0);
 
-  return retval;
+	return retval;
 }
 
 /*
@@ -82,9 +79,9 @@ static uint8_t bbspi_transfer(uint8_t cspin, uint8_t val)
  */
 void flash_init()
 {
-  gpio_out(GPIO_SPI_NCS, 1);
-  gpio_out(GPIO_SPI_SCLK, 0);
-  gpio_out(GPIO_SPI_MOSI, 0);
+	gpio_out(GPIO_SPI_NCS, 1);
+	gpio_out(GPIO_SPI_SCLK, 0);
+	gpio_out(GPIO_SPI_MOSI, 0);
 }
 
 /*
@@ -92,22 +89,21 @@ void flash_init()
  */
 int flash_write(uint32_t addr, uint8_t *buf, int count)
 {
-  int i;
+	int i;
 
-  bbspi_transfer(1,0);
-  bbspi_transfer(0,0x06);
-  bbspi_transfer(1,0);
-  bbspi_transfer(0,0x02);
-  bbspi_transfer(0,(addr & 0xFF0000) >> 16);
-  bbspi_transfer(0,(addr & 0xFF00) >> 8);
-  bbspi_transfer(0,(addr & 0xFF));
-  for ( i = 0; i < count; i++ )
-  {
-    bbspi_transfer(0,buf[i]);
-  }
-  bbspi_transfer(1,0);
+	bbspi_transfer(1,0);
+	bbspi_transfer(0,0x06);
+	bbspi_transfer(1,0);
+	bbspi_transfer(0,0x02);
+	bbspi_transfer(0,(addr & 0xFF0000) >> 16);
+	bbspi_transfer(0,(addr & 0xFF00) >> 8);
+	bbspi_transfer(0,(addr & 0xFF));
+	for ( i = 0; i < count; i++ ) {
+		bbspi_transfer(0,buf[i]);
+	}
+	bbspi_transfer(1,0);
 
-  return count;
+	return count;
 }
 
 /*
@@ -115,20 +111,19 @@ int flash_write(uint32_t addr, uint8_t *buf, int count)
  */
 int flash_read(uint32_t addr, uint8_t *buf, int count)
 {
-  int i;
-  bbspi_transfer(1,0);
-  bbspi_transfer(0,0x0b);
-  bbspi_transfer(0,(addr & 0xFF0000) >> 16);
-  bbspi_transfer(0,(addr & 0xFF00) >> 8);
-  bbspi_transfer(0,(addr & 0xFF));
-  bbspi_transfer(0,0);
-  for ( i = 0; i < count; i++ )
-  {
-    buf[i] = bbspi_transfer(0, 0);
-  }
-  bbspi_transfer(1,0);
+	int i;
+	bbspi_transfer(1,0);
+	bbspi_transfer(0,0x0b);
+	bbspi_transfer(0,(addr & 0xFF0000) >> 16);
+	bbspi_transfer(0,(addr & 0xFF00) >> 8);
+	bbspi_transfer(0,(addr & 0xFF));
+	bbspi_transfer(0,0);
+	for ( i = 0; i < count; i++ ) {
+		buf[i] = bbspi_transfer(0, 0);
+	}
+	bbspi_transfer(1,0);
 
-  return count;
+	return count;
 }
 
 /*
@@ -136,14 +131,14 @@ int flash_read(uint32_t addr, uint8_t *buf, int count)
  */
 void flash_serase(uint32_t addr)
 {
-  bbspi_transfer(1,0);
-  bbspi_transfer(0,0x06);
-  bbspi_transfer(1,0);
-  bbspi_transfer(0,0xD8);
-  bbspi_transfer(0,(addr & 0xFF0000) >> 16);
-  bbspi_transfer(0,(addr & 0xFF00) >> 8);
-  bbspi_transfer(0,(addr & 0xFF));
-  bbspi_transfer(1,0);
+	bbspi_transfer(1,0);
+	bbspi_transfer(0,0x06);
+	bbspi_transfer(1,0);
+	bbspi_transfer(0,0xD8);
+	bbspi_transfer(0,(addr & 0xFF0000) >> 16);
+	bbspi_transfer(0,(addr & 0xFF00) >> 8);
+	bbspi_transfer(0,(addr & 0xFF));
+	bbspi_transfer(1,0);
 }
 
 /*
@@ -164,12 +159,12 @@ flash_berase()
  */
 uint8_t flash_rsr()
 {
-  uint8_t retval;
-  bbspi_transfer(1,0);
-  bbspi_transfer(0,0x05);
-  retval = bbspi_transfer(0,0);
-  bbspi_transfer(1,0);
-  return retval;
+	uint8_t retval;
+	bbspi_transfer(1,0);
+	bbspi_transfer(0,0x05);
+	retval = bbspi_transfer(0,0);
+	bbspi_transfer(1,0);
+	return retval;
 }
 
 
@@ -225,14 +220,12 @@ int flash_sdb_check()
 
 	uint32_t entry_point[6] = {0x000000, 0x100, 0x200, 0x300, 0x170000, 0x2e0000};
 
-	for (i = 0; i < ARRAY_SIZE(entry_point); i++)
-	{
+	for (i = 0; i < ARRAY_SIZE(entry_point); i++) {
 		flash_read(entry_point[i], (uint8_t *)&magic, 4);
 		if (magic == SDB_MAGIC)
 			break;
 	}
-	if (magic == SDB_MAGIC)
-	{
+	if (magic == SDB_MAGIC) {
 		mprintf("Found SDB magic at address 0x%06X!\n", entry_point[i]);
 		wrc_sdb.drvdata = NULL;
 		wrc_sdb.entrypoint = entry_point[i];
